requirements.txt
pgoogle-generativeai
groq
beautifulsoup4
playwright
pyyaml
lxml
openai
google-generativeai
